--- src/app/modules/services/components/grua/grua.component.html
@@ -1,1 +1,0 @@
-<p>grua works!</p>
