--- src/app/modules/services/components/grua/grua.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-grua',
-  templateUrl: './grua.component.html',
-  styleUrls: ['./grua.component.scss']
-})
-export class GruaComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
