--- src/app/modules/services/components/maquina/maquina.component.html
@@ -1,1 +1,0 @@
-<p>maquina works!</p>
