--- src/app/modules/services/components/maquina/maquina.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { MaquinaComponent } from './maquina.component';
-
-describe('MaquinaComponent', () => {
-  let component: MaquinaComponent;
-  let fixture: ComponentFixture<MaquinaComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ MaquinaComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(MaquinaComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
