--- src/app/modules/services/components/maquina/maquina.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-maquina',
-  templateUrl: './maquina.component.html',
-  styleUrls: ['./maquina.component.scss']
-})
-export class MaquinaComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
