+++ src/app/modules/services/components/table-services/table-services.component.html
@@ -1,11 +1,3 @@
-<!-- <div>
-  <h3>Hola</h3>
-  <ng-container *ngFor="let negociacion of negociaciones">
-    {{negociacion.obra}}
-
-  </ng-container>
-</div> -->
-
 <div class="container mt-5">
   <h3 class="text-center mb-5 titulo">Programaciones a Enviar</h3>
   <div class="table-responsive">
@@ -13,20 +5,20 @@
       <thead>
         <tr>
           <th scope="col">Obra</th>
-          <th scope="col">Material</th>
-          <th scope="col">Placa</th>
+          <th scope="col" *ngIf="this.path !== servicesEnum.grua">Material</th>
+          <th scope="col" *ngIf="this.path !== servicesEnum.maquina">Placa</th>
           <th scope="col">Origen</th>
-          <th scope="col">Destino</th>
-          <th scope="col">Acción</th>
+          <th scope="col" *ngIf="this.path !== servicesEnum.maquina">Destino</th>
+          <th scope="col" *ngIf="this.path !== servicesEnum.maquina">Acción</th>
         </tr>
       </thead>
       <tbody *ngFor="let negociacion of negociaciones">
         <tr>
           <td class="text-nowrap" scope="row">{{negociacion.obra}}</td>
-          <td class="text-nowrap">{{negociacion.material}}</td>
-          <td class="text-nowrap">{{negociacion.placa}}</td>
-          <td class="text-nowrap">{{negociacion.origen}}</td>
-          <td class="text-nowrap">{{negociacion.destino}}</td>
+          <td class="text-nowrap" *ngIf="this.path !== servicesEnum.grua">{{negociacion.material}}</td>
+          <td class="text-nowrap" *ngIf="this.path !== servicesEnum.maquina">{{negociacion.placa}}</td>
+          <td class="text-nowrap" *ngIf="this.path !== servicesEnum.maquina">{{negociacion.origen}}</td>
+          <td class="text-nowrap" *ngIf="this.path !== servicesEnum.maquina">{{negociacion.destino}}</td>
           <td>
             <div class="row mx-0 px-0">
               <button (click)="eliminarProgramacion(negociacion.customId)" class="btn btn-outline-danger col-auto">
