+++ src/app/modules/services/services.module.ts
@@ -2,9 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ServicesRoutingModule } from './services-routing.module';
-import { VolquetaComponent } from './components/volqueta/volqueta.component';
-import { GruaComponent } from './components/grua/grua.component';
-import { MaquinaComponent } from './components/maquina/maquina.component';
+import { ServiceComponent } from './components/service/service.component';
 import { ServiceLayoutComponent } from './components/service-layout/service-layout.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {CoreModule} from "../core/core.module";
@@ -13,9 +11,7 @@ import { TableServicesComponent } from './components/table-services/table-servic
 
 @NgModule({
   declarations: [
-    VolquetaComponent,
-    GruaComponent,
-    MaquinaComponent,
+    ServiceComponent,
     ServiceLayoutComponent,
     TableServicesComponent
   ],
@@ -24,7 +20,7 @@ import { TableServicesComponent } from './components/table-services/table-servic
     ServicesRoutingModule,
     FormsModule,
     ReactiveFormsModule,
-    CoreModule
+    CoreModule,
   ]
 })
 export class ServicesModule { }
